pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyairtel"
version = "0.1.0"
description = "Python client for Airtel Money Tanzania — Collection & Disbursement APIs"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Ronald Isack Gosso", email = "ronaldgosso@gmail.com" }
]
keywords = ["airtel", "airtel-money", "tanzania", "mobile-money", "fintech", "payments"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "requests>=2.28.0",
]

[project.optional-dependencies]
encryption = ["pycryptodome>=3.18.0"]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.10",
    "responses>=0.23",
    "ruff>=0.4.0",
    "mypy>=1.0",
    "black>=24.0",
    "types-requests",
]

[project.urls]
Homepage = "https://ronaldgosso.github.io/pyairtel"
Repository = "https://github.com/ronaldgosso/pyairtel"
"Bug Tracker" = "https://github.com/ronaldgosso/pyairtel/issues"

[tool.hatch.build.targets.wheel]
packages = ["pyairtel"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
strict = false
ignore_missing_imports = true
disallow_untyped_defs = true
warn_return_any = true

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.pytest.ini_options]
testpaths = ["tests"]
